Moved the common expenses manager components to its own folder
