+++ src/components/common/ExpensesManager/CategorySelector.js
