+++ src/components/common/ExpensesManager/CategorySelector.test.js
