+++ src/components/common/ExpensesManager/Summaries/EntriesSummary.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { getSumFromEntries } from '../../../helpers/entriesHelper';
+import { getSumFromEntries } from '../../../../helpers/entriesHelper';
 
 function GetEntriesList(entries) {
   return entries.map((entry, key) => <li key={key}>{entry.ammount} {entry.description} {entry.category}</li>);
