+++ src/components/common/ExpensesManager/Summaries/EntrySummaryWithFilter.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import EntriesSummary from './EntriesSummary';
 import CategorySelector from '../CategorySelector';
 import { connect } from 'react-redux';
-import { categoryChange } from '../../../redux/expensesManager/actionCreators';
-import { getEntryCategoryOption } from '../../../helpers/entriesHelper'; 
+import { categoryChange } from '../../../../redux/expensesManager/actionCreators';
+import { getEntryCategoryOption } from '../../../../helpers/entriesHelper'; 
 
 class EntrySummaryWithFilter extends Component {
   constructor(props) {
