+++ src/components/common/ExpensesManager/Summaries/EntrySummaryWithFilter.test.js
